Made it easier to download enron test dataset
