+++ gpt_neox/__init__.py
@@ -1,6 +1,6 @@
 from gpt_neox.autoregressive_wrapper import AutoregressiveWrapper
 from gpt_neox.data_utils import get_tokenizer, read_enwik8_data
-from gpt_neox.datasets import TextSamplerDataset, GPT2Dataset
+from gpt_neox.datasets import TextSamplerDataset, TFRecordDataset,JsonShardedDataset
 from gpt_neox.gpt_neox import GPTNeoX
 from gpt_neox.utils import *
 from gpt_neox.data_downloader_registry import prepare_data
